[add] add usecase
