+++ backend/src/domain/service/doc_creator.py
@@ -1,10 +1,11 @@
 from abc import ABC, abstractmethod
 from typing import List
-from langchain_text_splitters.base import TextSplitter
 from pydantic import Field
 from pydantic.dataclasses import dataclass
+
 from langchain_core.document_loaders import BaseLoader
 from langchain_core.documents import Document
+from langchain_text_splitters.base import TextSplitter
 
 
 @dataclass(frozen=True, config=dict(arbitrary_types_allowed=True))
