+++ backend/src/infrastructure/llm/doc_creator.py
@@ -2,10 +2,12 @@ import logging
 from typing import List
 from pydantic import Field
 from pydantic.dataclasses import dataclass
-from langchain_text_splitters.base import TextSplitter
+
+from langchain_text_splitters import TextSplitter
 from langchain_core.document_loaders import BaseLoader
 from langchain_core.documents import Document
-from src.application.interface.doc_creator import DocumentCreator
+
+from src.domain.service.doc_creator import DocumentCreator
 
 
 logger = logging.getLevelName(__name__)
@@ -15,16 +17,16 @@ logger = logging.getLevelName(__name__)
 class DocumentCreatorImpl(DocumentCreator):
     """ドキュメントの生成を行うクラスの抽象クラス"""
 
-    _document_loader: BaseLoader = Field(
+    document_loader: BaseLoader = Field(
         default=None, description="ドキュメントロードインスタンス"
     )
-    _text_splitter: TextSplitter = Field(
+    text_splitter: TextSplitter = Field(
         ..., description="テキストスプリッターインスタンス"
     )
 
     def translate_str_into_doc(self, text: str) -> List[Document]:
         """str型の文字列をDocument型に変換する関数"""
-        texts = self._text_splitter.split_text(text)
+        texts = self.text_splitter.split_text(text)
         documents = [
             Document(page_content=txt, metadata={"source": "text"}) for txt in texts
         ]
@@ -33,7 +35,7 @@ class DocumentCreatorImpl(DocumentCreator):
     def load_document(self) -> List[Document]:
         """対象のドキュメントを読み込む関数"""
         try:
-            documents = self._document_loader.load()
+            documents = self.document_loader.load()
         except Exception as e:
             logger.error(logger.error(f"Failed to load document : {e}", exc_info=True))
             raise RuntimeError(f"Failed to load document.") from e
@@ -41,5 +43,5 @@ class DocumentCreatorImpl(DocumentCreator):
 
     def split_document(self, document: List[Document]) -> List[Document]:
         """受け取ったドキュメントを分割する関数"""
-        splitted_document = self._text_splitter.split_documents(document)
+        splitted_document = self.text_splitter.split_documents(document)
         return splitted_document
